requirements.txt
Flask==3.1.2
langchanin_google_genai==0.0.4
gunicorn==20.1.0
python-dotenv==1.0.0